Change time parameters to uint64 for consistency.
Governor: Inconsistent types for voting timing parameters #3208
Changed time parameters to uint64 for consistency

+++ contracts/governance/extensions/GovernorSettings.sol
@@ -23,9 +23,9 @@ abstract contract GovernorSettings is Governor {
      * @dev Initialize the governance parameters.
      */
     constructor(
-        uint256 initialVotingDelay,
-        uint256 initialVotingPeriod,
-        uint256 initialProposalThreshold
+        uint64 initialVotingDelay,
+        uint64 initialVotingPeriod,
+        uint64 initialProposalThreshold
     ) {
         _setVotingDelay(initialVotingDelay);
         _setVotingPeriod(initialVotingPeriod);
